+++ src/components/player/layout/PlayControl.tsx
